LocusZoom

LocusZoom

LocusZoom(
  gwas_results,
  range = 5e+05,
  position_zoom = NULL,
  use_biomaRt = TRUE,
  TxDb_version = c("hg18", "hg19"),
  GRCh = NULL,
  pvalue = "p.value",
  chromosome = "chr",
  position = "pos",
  rs = "rs",
  range_filter = NULL,
  draw_genes = TRUE,
  pval_threshold = 5
)

Arguments

gwas_results

data.frame With atleast three columns: Chromosome, position and p-value

range

numeric (default 5e+05) Range of the plot in genome positions.

position_zoom

numeric (default NULL) Center position of the plot if supplied. If NULL, the center position is the lowerst p-value of the gwas_results table.

use_biomaRt

bool (default TRUE) Set to TRUE to use biomaRt to retrieve the annotation of the genes. If set to FALSE TxDb.Hsapiens.UCSC.hgXX.knownGene will be used. Use the argument TxDb_version to select between hg19 or hg18. Note that when using TxDb the correspondent packages needs to be installed, providing a internet-free annotation option.

TxDb_version

character (default hg18) TxDb version to use. This argument is only used when use_biomaRt is set to FALSE.

GRCh

numeric (default NULL) GRCh version to connect to if not the current GRCh38, currently this can only be 37, if is different from 37 the GRCh38 version will be used. (Parameter passed to biomaRt::useEnsembl). This argument is only used when use_biomaRt is set to TRUE.

pvalue

character (default "p.value") Column name of gwas_results that contains the p-values.

chromosome

character (default "chr") Column name of gwas_results that contains the chromosome names.

position

character (default "pos") Column name of gwas_results that contains the positions.

rs

character (default "rs") Column name of gwas_results that contains the rs ID of the SNPs. This column will only be used if pval_threshold is different than NULL, therefore there is no need of supplying this argument if there is no interest to plot the labels of the rs IDs.

range_filter

numeric (default NULL). If not NULL, only the annotations with a range larger than range_filter will be displayed.

draw_genes

bool (default TRUE). Select if the genes of the region plot are to be plotted below the p-values plot.

pval_threshold

numeric (default 5) Threshold of -log_10p-value to draw labels with the SNP IDs. If NULL no labels will be drawn.

Value