Allelic frequency

Calculates the frequency of the A allele on a server side GenotypeData object.

ds.alleleFrequency(
  genoData,
  type = "combined",
  method = "fast",
  snpBlock = 5000L,
  datasources = NULL
)

Arguments

genoData

character Name of the GenotypeData object on the server

type

character (default "combined") Type of analysis, if ("split"), the frequencies will be calculated for each study server. If ("combined") a pooled methodology will be performed.

method

character (default "fast") If "fast" an optimized fast method will be used; if "slow" the function GWASTools::alleleFrequency will be used, this is notably slower.

snpBlock

integer (default 5000L) number of SNPs to read at each iteration, tune this parameter to improve performance. This argument is only used when method = "fast"

datasources

a list of DSConnection-class objects obtained after login.

Value

A data frame with a row for each SNP. Columns "M" for males, "F" for females, and "all" for all scans give frequencies of the A allele. Sample size for males, females, and all is returned as "n.M", "n.F", and "n", respectively. "MAF" is the minor allele frequency over all scans.

Details

(From alleleFrequency documentation): Counts male heterozygotes on the X and Y chromosomes as missing values, and any genotype for females on the Y chromosome as missing values. A "sex" variable must be present in the scan annotation slot of genoData. Samples with missing sex are included in the allele counts for "all" and "MAF" for autosomes, but not for sex chromosomes.