Server-side Differential Gene Expression analysis using limma

This function performs a non-disclosive Differential Gene Expression Analysis using limma package from Bioconductor.

ds.limma(
  model,
  Set,
  type.data = "microarray",
  contrasts = NULL,
  levels = "design",
  coef = 2,
  sva = FALSE,
  annotCols = NULL,
  method = "ls",
  robust = FALSE,
  normalization = "none",
  voomQualityWeights = FALSE,
  big = FALSE,
  datasources = NULL
)

Arguments

model

formula indicating the condition (left side) and other covariates to be adjusted for (i.e. condition ~ covar1 + ... + covar2). The fitted model is: feature ~ condition + covar1 + ... + covarN

Set

name of the DataSHIELD object to which the ExpresionSet or RangedSummarizedExperiment has been assigned

type.data

optional parameter that allows the user to specify the number of CPU cores to use during

sva

logical value

annotCols

the column names of the annotation available in the ExpresionSet or RangedSummarizedExperiment (see fData() function)

method

String indicating the method used in the regression: "ls" or "robust". (Default: "ls")

robust

Logical value indicating whether robust method is applied in the eBayes function of limma. Default is FALSE.

normalization

String indicating the normalize method used when using voom for RNAseq data (see normalized.method argument in limma::vomm for possible values) #' @param voomQualityWeights Logical value indicating whether limma::voomWithQualityWeights should be used instead of limma::voom. Default is FALSE and hence the pipeline uses limma::voom to transform RNAseq data.

datasources

a list of DSConnection-class objects obtained after login. If the datasources argument is not specified the default set of connections will be used: see datashield.connections_default.

Details

Implementation of Bioconductor's limma in DataSHIELD using MEAL package

Examples